fix: Correct MonoType::apply for instantiation use
We assume that a variable that were replaced are the actual root so we
should not recurse further. If we do not have this guard `instantiate`
breaks as it may generate a substitution map like `{ A => B, B => C }`
which would replace `A` with the (fresh) variable `B` which would then
be replaced again due to `B => C` which is very wrong (`B` != `fresh B`).